refactor TaskStorage out
